Fix transparent pixel encoding
The code was treating transparent pixels as if the mask channel were an alpha channel, but unfortunately the Arduboy2 Sprites format requires that both a mask bit and its corresponding image bit must be zero for the pixel to be truly transparent. This change fixes the issue by ensuring that if a source pixel is transparent the image data isn't encoded, leaving the destination bit as zero.